pyproject.toml
[project]
name = "final_project"
version = "0.1.0"
description = ""
authors = [
    {name = "Sophia Shen",email = "sophiashen767@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy (>=2.3.4,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "fastapi[standard] (>=0.121.1,<0.122.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
